NahelTeam/nahl-backend/requirements.txt
Flask>=2.0
Flask-Cors
python-dotenv
Pillow
